Refactored ToggleSection component into DasboardToggleSection, ToggleSection and NoHighlightToggleSection
