+++ src/client/components/ToggleSection/index.jsx
@@ -1,28 +1,7 @@
-import multiInstance from '../../utils/multiinstance'
-import { TOGGLE_SECTION__TOGGLE } from '../../actions'
-
+import DashboardToggleSection from './DashboardToggleSection'
+import NoHighlightToggleSection from './NoHighlightToggleSection'
 import ToggleSection from './ToggleSection'
 
-export default multiInstance({
-  name: 'ToggleSection',
-  actionPattern: 'TOGGLE_SECTION__',
-  dispatchToProps: (dispatch) => ({
-    open: (isOpen) =>
-      dispatch({
-        type: TOGGLE_SECTION__TOGGLE,
-        isOpen,
-      }),
-  }),
-  component: ToggleSection,
-  reducer: (state = {}, { type, isOpen }) => {
-    switch (type) {
-      case TOGGLE_SECTION__TOGGLE:
-        return {
-          ...state,
-          isOpen,
-        }
-      default:
-        return state
-    }
-  },
-})
+export { DashboardToggleSection, NoHighlightToggleSection, ToggleSection }
+
+export default ToggleSection
